Don't emit multiple events for the same resource_id
The output of the compose command that is used for parsing events can contain multiple events for containers with health checks.  With one event status Waiting then a following event with the status Healthy. This causes the module to erroneously report changes.

This change ensures only the last event for each container is reported.